Keep state of each trace in a hash map (#4)
* Keep state of each trace in a hash map

Push to a span stack when a span is started and pop when the span is
completed.  Keep track of each completed span and when the stack is
empty (all spans have been ended), send the whole vector to datadog.

The span stacks are blocked by an RW lock and are also
mutable/consuming, so memory should not grow during operation (completed
spans are removed from the hashmap, and are freed).