Add logging to tracer
This will print out the "current" trace ID and span ID for every log
line if the LoggingConfig is passed into the creation of the
DatadogTracer.  The actual logging is in the looping receiver, so it
will not take up time on the thread which is calling "debug!, info!,
etc."

The logger also incorporates parts of the FilterLogger, allowing a
filter for log lines based on module or body containing a string (filter
strings can be passe din a vector to the LoggingConfig).

The current span's trace ID and span ID is kept current and is used to
print out the logs.  If they are not set (all spans were popped or
haven't started), then a normal log line will print out.

Co-authored-by: Kareem Sait <kareem.sait@paidy.com>