Major refactoring to support tokio/rs-tracing
Removed manual activation of tracing and instead use rust tracing events
to instrument and send datadog tracing evens as a subscriber.

Also reimplemented log with support for trace-id and span-id.

Use thread local storage to account for current trace and span IDs so
that new spans get the appropriate parents.