Automatically push trace parent span
Now it is no longer necessary to decide on the parent; it will
automatically be pushed when a new trace is started (first span where
span collection is empty for the given trace ID).  This parent wil be
re-calculated if all events are sent on a trace (because the current
span collection for that trace will be consumed and further create
events will create a whole new span collection, including a new parent.

Now, draining a span collection consumes it, so it can no longer be used
once the "send_trace" event is received for a trace.