Overhaul of trace ID identification and sending (#13)
* Overhaul of trace ID identification and sending

Trace ID cannot be identified per thread, because it is possible for a
thread to service many traces, if there is any work-stealing amongst the
threads (if one "task" running a trace awaits, the thread could
suspend that task and grab another, starting a new trace).

Since it is impossible for the datadog tracer to know the trace ID in
this circumstance, we must force the user to tell us the ID through the
span attributes (key/value pairs in the span! macro call).  Likewise for
the event to send to datadog, as this even can occur outside of any of
the spans for the trace the user wishes to send (i.e. once the trace is
complete, the user asks for the tradce to be shipped off to datadog;
this is a common use case).

Once the user qualifies the trace ID for a new span, we can internally
keep track of which trace each thread is currently servicing based off
of the "enter/exit" span calls (one thread can only "ENTER" a span one
at a time; they cannot ever be "in" two spans at once).  Keeping track
of which threads are in which traces (and vice versa), any span
operations and/or log operations can look up the "current" span based on
the thread ID alone.

Furthmore, add support for APM analytics by adding an "APM Config"
option to the main config.  Turning on the apm_config (using apm_enabled
flag) will send APM Analytics to datadog, allowing APM metrics to be
tallied and used (these get set in the tags of each span).

* move tokio back to dev-deps

* cleanup unneeded file