Add unique ID generator for u64
Also change how thread IDs are generated to just using a static atomic
counter, starting from 0 (32-bit).

Unique IDs are generated by using 48-bits (6 bytes) as the number of
milliseconds since epoch and a 16-bit static counter, allowing for 65k
uniqueIDs to be generated in one second.

This unique ID generator is meant to be used to generate trace IDs.

Span IDs are still generated with timestamp_nanos because they just have
to be unique per trace and not be the same as the traceID.