Add functions to get current Span ID and sampling rate (#21)
The sampling rate is easy, since it's globally set when the tracer is
installed as the global default (only happens once).

The span ID is tougher, because it changes as a thread enters and leaves
spans (even between different traces!).  But still, a thread can only be
in one span at a time, so thread_local span ID can be used to always
keep the most current "entered" span, much as the tracer thread does
with its "span collection".  The get current span function, though, must
be available from teh same thread that calls the span! and event!
macros, because the processor thread is running in the background and
can't return that info to the actual running program threads.  This
means we have to keep the span ID in synch with what the background
thread knows.