Fix memory leak in span management
Every time a new span was created, a span id -> trace id mapping was
made to be able to figure out which trace needed to be updated when a
span was exited by a thread or ended.  However, this map was never
cleaned up, allowing it to grow indefinitely.

Fix is to remove the span id -> trace id mapping from the map once the
span is ended, as it would no longer be needed after tha tpoint (a
thread cannot exit a span if the span is ended and doesn't exist, after
all).

Another slight convern was how the system maps thread IDs to the current
trace IDs.  This mapping is necessary since threads may enter
and exit spans, depending on future execution, and different threads may
be called upon to service the same trace.  The same thread may also
service multiple traces at the same time.  Because of this, a map is
kept of which thread is currently operating on which trace (and vice
versa).  However, this map only has a single mapping from thread to
trace (and viceversa), meaning if a thread operated on a different
trace, the original mapping would be lost.  This is a problem because we
use that mapping to also remvoe the entry from the reverse lookup map,
and if that entry is erased, we cannot remove a potentially stale entry
from the trace -> thread map.  Since trace ids are always uniquely
generated, there is no chance that an existing entry would be
overwritten and cleared at any point in the future, resulting in a leak.

The solution was to just remove the mapping (forward and backward) when
a thread exits a span.  The only risk is if a thread logs or tries
to send a trace when it never entered a span.  This can only happen if
thread A exits a span and then thread B tries to log on that trace.
Theoretically, this shouldn't be possible, since "enter span" has to be
called in some way for the system to be able to end a span (and hence
send a trace), and any log has to be in a span as well, so the same rule
should also apply.

Therefore, this change should be safe and should ensure a leak doesn't
occur in the off chance that a thread swaps out mid-trace due to
async/await mechanics.