Increase http client send performance
Use crossbeam to defer the sending of the spans to threads running in
the background, rather than blocking the handler thread.  The handler
thread must remain a single-threaded, serial loop to guarantee the
ordering of all the different spans, otherwise they could get
out-of-order.  This means that any hiccup in this thread will slow down
the processing, so that the inflow of new requests could back up.  If
this is the normal case (where inflow > outflow), the channel will
necessarily back up, causing more and more memory to be used.

To help unblock the handler channel so that the outflow can be
maximized, any I/O should be offloaded to other threads.  This has the
advantage of allowing more than one thread to handle certain operations,
since the ordering has already been set.  In the case of the HTTP client
send, the spans have all been collated and the trace has been closed,
meaning the order is now moot and the various event sends can be handled
in parallel with no fear of misordering.  This means the time-intensive
HTTP call can be farmed out to a number of helper threads, whereas
before, the main thread was doing the work each time in a blocking
fashion.

This has seen a remarkable performance increase, although there still
remains issues when using the standard logger, as the span/trace ID
lookups (to put into the logging message) are too slow.  However, the
I/O to the datadog agent via HTTP should now no longer be a blocker for
handling events, thus increasing the throughput of the channels
immensely.

This was accomplished by spawning a number of helper threads in
DdAgentClient along with cloned multi-sender, multi-consumer channels
(courtesy of the crossbeam-channel crate), which handle high-speed
locking and synchronization between the various receivers.  The "send()"
command (run when the "send_trace" tag is used in an event! command) now
only pushes the Span list onto this mpmc channel, and continues on.  The
threads that start up when the DdAgentClient is created handle the
actual reading of this shared channel and processing of the client calls
via HTTP.

The number of these helper threads defaults to 4, but can be set in the
Config of the DatadogTracing::new() call.