Fix logging filter
The log filter was changed to improve perforamnce, but it unfortunately
became a direct equality  check (which is rarely true) instead of a
standard "contains" check.

This has been fixed to now check for a contains on the mod filter versus
the mod line in the log event.

A test has been added, but there is no way to assert, since this is a
stdout call (from a different thread, so "gag" crate won't work).
Merely run the test and check the output manually to see that the lines
do NOT print out.